+++ Sources/GoodNetworking/Logging/DataTaskLogging.swift
@@ -1,5 +1,5 @@
 //
-//  LoggingEventMonitor.swift
+//  DataTaskLogging.swift
 //  GoodNetworking
 //
 //  Created by Matus Klasovity on 30/01/2024.
@@ -10,11 +10,11 @@ import Foundation
 
 // MARK: - Logging
 
-@NetworkActor internal extension DataTaskProxy {
+internal extension DataTaskProxy {
 
-    private static var maxLogSizeBytes: Int { 32_768 } // 32 kB
+    @NetworkActor private static var maxLogSizeBytes: Int { 32_768 } // 32 kB
 
-    internal func prepareRequestInfo() -> String {
+    @NetworkActor func prepareRequestInfo() -> String {
         """
         🚀 \(task.currentRequest?.method.rawValue ?? "-") \(task.currentRequest?.url?.absoluteString ?? "<unknown>")
         \(prepareResponseStatus(response: task.response, error: receivedError))
@@ -30,7 +30,7 @@ import Foundation
         """
     }
 
-    private func prepareResponseStatus(response: URLResponse?, error: (any Error)?) -> String {
+    @NetworkActor private func prepareResponseStatus(response: URLResponse?, error: (any Error)?) -> String {
         guard let response = response as? HTTPURLResponse else { return "" }
         let statusCode = response.statusCode
 
@@ -38,31 +38,37 @@ import Foundation
         logMessage.append(HTTPURLResponse.localizedString(forStatusCode: statusCode))
 
         if error != nil {
-            logMessage.append("\n🚨 Error: \(error?.localizedDescription)")
+            logMessage.append("\n🚨 Error: \(error?.localizedDescription ?? "<nil>")")
         }
 
         return logMessage
     }
 
-    private func prepareHeaders(request: URLRequest?) -> String {
+    @NetworkActor private func prepareHeaders(request: URLRequest?) -> String {
         guard let request, let headerFields = request.allHTTPHeaderFields else { return " <no headers>" }
 
         return headerFields.map { key, value in
-            " - \(key): \(value ?? "<nil>")"
+            " - \(key): \(value)"
         }
         .joined(separator: "\n")
     }
 
-    private func prettyPrintMessage(data: Data?) -> String {
+    @NetworkActor private func prettyPrintMessage(data: Data?) -> String {
         guard let data else { return "" }
 
         guard data.count < Self.maxLogSizeBytes else {
             return "💡 Data size is too big (\(data.count) bytes), console limit is \(Self.maxLogSizeBytes) bytes"
         }
 
+        let serializationOptions: JSONSerialization.WritingOptions = if #available(macOS 10.15, *) {
+            [.prettyPrinted, .withoutEscapingSlashes]
+        } else {
+            [.prettyPrinted]
+        }
+        
         if let string = String(data: data, encoding: .utf8) {
             if let jsonData = try? JSONSerialization.jsonObject(with: data, options: []),
-               let prettyPrintedData = try? JSONSerialization.data(withJSONObject: jsonData, options: [.prettyPrinted, .withoutEscapingSlashes]),
+               let prettyPrintedData = try? JSONSerialization.data(withJSONObject: jsonData, options: serializationOptions),
                let prettyPrintedString = String(data: prettyPrintedData, encoding: .utf8) {
                 return prettyPrintedString
             } else {
