+++ Sources/GoodNetworking/Resource/Placeholdable.swift
