+++ Sources/GoodNetworking/Resource/RemoteResource.swift
