+++ Sources/GoodNetworking/Resource/ResourceOperations.swift
