+++ Sources/GoodNetworking/Resource/ResourceState.swift
