+++ Sources/GoodNetworking/Session/NetworkSession.swift
@@ -1,317 +1,425 @@
-////
-////  NetworkSession.swift
-////  GoodNetworking
-////
-////  Created by Dominik Pethö on 8/17/20.
-////
 //
-//import Foundation
+//  GRSession.swift
+//  GoodNetworking
 //
-///// Executes network requests for the client app.
-/////
-///// `NetworkSession` is responsible for sending, downloading, and uploading data through a network session.
-///// It uses a base URL provider and a session provider to manage the configuration and ensure the session's validity.
-//public actor OLD_NetworkSession: Hashable {
+//  Created by Filip Šašala on 02/07/2025.
 //
-//    public static func == (lhs: OLD_NetworkSession, rhs: OLD_NetworkSession) -> Bool {
-//        lhs.sessionId == rhs.sessionId
+
+import Foundation
+
+// MARK: - Initialization
+
+@NetworkActor public final class NetworkSession: NSObject, Sendable {
+    
+    nonisolated public let name: String
+
+    private let baseUrl: any URLConvertible
+    private let sessionHeaders: HTTPHeaders
+    private let interceptor: any Interceptor
+    private let logger: any NetworkLogger
+
+    private let configuration: URLSessionConfiguration
+    private let delegateQueue: OperationQueue
+    private lazy var session: URLSession = {
+        URLSession(
+            configuration: configuration,
+            delegate: NetworkSessionDelegate(for: self),
+            delegateQueue: delegateQueue
+        )
+    }()
+    
+    /// Holds references to `DataTaskProxy` objects based on
+    /// DataTask `taskIdentifier`-s.
+    private var activeTasks: [Int: DataTaskProxy] = [:]
+
+    nonisolated public init(
+        baseUrl: any URLConvertible,
+        baseHeaders: HTTPHeaders = [],
+        interceptor: any Interceptor = NoInterceptor(),
+        logger: any NetworkLogger = PrintNetworkLogger(),
+        name: String? = nil
+    ) {
+        self.baseUrl = baseUrl
+        self.sessionHeaders = baseHeaders
+        self.interceptor = interceptor
+        self.logger = logger
+        self.name = name ?? "NetworkSession"
+
+        let operationQueue = OperationQueue()
+        operationQueue.name = "NetworkActorSerialExecutorOperationQueue"
+        operationQueue.underlyingQueue = NetworkActor.queue
+
+        let configuration = URLSessionConfiguration.ephemeral
+        configuration.httpAdditionalHeaders = baseHeaders.map { $0.resolveHeader() }.reduce(into: [:], { $0[$1.name] = $1.value })
+
+        self.configuration = configuration
+        self.delegateQueue = operationQueue
+
+        // create URLSession lazily, isolated on @NetworkActor, when requested first time
+
+        super.init()
+    }
+
+}
+
+internal extension NetworkSession {
+
+    func proxyForTask(_ task: URLSessionTask) -> DataTaskProxy {
+        if let existingProxy = self.activeTasks[task.taskIdentifier] {
+            return existingProxy
+        } else {
+            let newProxy = DataTaskProxy(task: task, logger: logger)
+            self.activeTasks[task.taskIdentifier] = newProxy
+            return newProxy
+        }
+    }
+    
+    func closeProxyForTask(_ task: URLSessionTask) {
+        self.activeTasks.removeValue(forKey: task.taskIdentifier)
+    }
+
+}
+
+// MARK: - Network session delegate
+
+final class NetworkSessionDelegate: NSObject {
+
+    private unowned let networkSession: NetworkSession
+
+    internal init(for networkSession: NetworkSession) {
+        self.networkSession = networkSession
+    }
+
+}
+
+extension NetworkSessionDelegate: URLSessionDelegate {
+
+    public func urlSession(
+        _ session: URLSession,
+        didReceive challenge: URLAuthenticationChallenge,
+        completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void
+    ) {
+#warning("TODO: Implement SSL pinning/certificate validation")
+        completionHandler(.performDefaultHandling, nil)
+    }
+
+}
+
+extension NetworkSessionDelegate: URLSessionTaskDelegate {
+
+    func urlSession(
+        _ session: URLSession,
+        task: URLSessionTask,
+        didCompleteWithError error: (any Error)?
+    ) {
+        NetworkActor.assumeIsolated {
+            networkSession
+                .proxyForTask(task)
+                .dataTaskDidComplete(withError: error)
+        }
+    }
+
+}
+
+extension NetworkSessionDelegate: URLSessionDataDelegate {
+
+    func urlSession(
+        _ session: URLSession,
+        dataTask: URLSessionDataTask,
+        didReceive response: URLResponse
+    ) async -> URLSession.ResponseDisposition {
+        return .allow
+    }
+
+    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
+        NetworkActor.assumeIsolated {
+            networkSession
+                .proxyForTask(dataTask)
+                .dataTaskDidReceive(data: data)
+        }
+    }
+
+}
+
+// MARK: - Request
+
+extension NetworkSession {
+    
+    // MARK: Result
+    
+    public func requestResult<T: Decodable, F: Error>(
+        endpoint: Endpoint,
+        validationProvider: any ValidationProviding<F> = DefaultValidationProvider()
+    ) async -> Result<T, F> {
+        do {
+            let response: T = try await request(endpoint: endpoint, validationProvider: validationProvider)
+            return .success(response)
+        } catch let error {
+            return .failure(error)
+        }
+    }
+    
+    // MARK: Codable
+
+    public func request<T: Decodable, F: Error>(
+        endpoint: Endpoint,
+        validationProvider: any ValidationProviding<F> = DefaultValidationProvider()
+    ) async throws(F) -> T {
+        do {
+            return try await request(endpoint: endpoint)
+        } catch let error {
+            throw validationProvider.transformError(error)
+        }
+    }
+
+    public func request<T: Decodable>(endpoint: Endpoint) async throws(NetworkError) -> T {
+        let data = try await request(endpoint: endpoint) as Data
+
+        // handle decoding corner cases
+        var decoder = JSONDecoder()
+        switch T.self {
+        case is Data.Type:
+            return data as! T
+            
+        case let t as WithCustomDecoder:
+            decoder = type(of: t).decoder
+            
+        default:
+            break
+        }
+
+        // decode
+        do {
+            let model = try decoder.decode(T.self, from: data)
+            return model
+        } catch let error as DecodingError {
+            throw error.asNetworkError()
+        } catch {
+            throw URLError(.cannotDecodeRawData).asNetworkError()
+        }
+    }
+    
+    // MARK: JSON
+    
+    @_disfavoredOverload
+    public func request(endpoint: Endpoint) async throws(NetworkError) -> JSON {
+        let responseData = try await request(endpoint: endpoint) as Data
+        guard let json = try? JSON(data: responseData) else {
+            throw URLError(.cannotDecodeRawData).asNetworkError()
+        }
+        return json
+    }
+    
+    // MARK: Raw
+    
+    @discardableResult
+    public func request(endpoint: Endpoint) async throws(NetworkError) -> Data {
+        guard let basePath = await baseUrl.resolveUrl()?.absoluteString,
+              let url = await endpoint.url(on: basePath)
+        else {
+            throw URLError(.badURL).asNetworkError()
+        }
+
+        // url + method
+        var request = URLRequest(url: url)
+        request.httpMethod = endpoint.method.rawValue
+        
+        // headers
+        endpoint.headers?.forEach { header in
+            request.setValue(header.value, forHTTPHeaderField: header.name)
+        }
+        
+        // encoding
+        switch endpoint.parameters {
+        case .parameters(let parameters):
+            if endpoint.encoding is URLEncoding {
+                applyQueryParameters(to: &request, endpoint, url)
+            } else if endpoint.encoding is JSONEncoding {
+                logger.logNetworkEvent(
+                    message: "Attempt to encode dictionary as body, use JSON or Encodable instead",
+                    level: .error,
+                    file: #file,
+                    line: #line
+                )
+                request.httpBody = try endpoint.parameters?.data()
+            } else if endpoint.encoding is AutomaticEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            }
+            
+        case .query(let queryItems):
+            if endpoint.encoding is URLEncoding {
+                applyQueryParameters(to: &request, endpoint, url)
+            } else if endpoint.encoding is JSONEncoding {
+                preconditionFailure("Attempt to encode query parameters as JSON body")
+            } else if endpoint.encoding is AutomaticEncoding {
+                applyQueryParameters(to: &request, endpoint, url)
+            }
+            
+        case .model(let encodableModel):
+            if endpoint.encoding is URLEncoding {
+                applyQueryParameters(to: &request, endpoint, url)
+            } else if endpoint.encoding is JSONEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            } else if endpoint.encoding is AutomaticEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            }
+            
+        case .data(let data):
+            if endpoint.encoding is URLEncoding {
+                preconditionFailure("Encoding raw data into query is not supported")
+            } else if endpoint.encoding is JSONEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            } else if endpoint.encoding is AutomaticEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            }
+            
+        case .json(let json):
+            if endpoint.encoding is URLEncoding {
+                applyQueryParameters(to: &request, endpoint, url)
+            } else if endpoint.encoding is JSONEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            } else if endpoint.encoding is AutomaticEncoding {
+                request.httpBody = try endpoint.parameters?.data()
+            }
+            
+        case .none:
+            request.httpBody = nil
+        }
+
+        return try await executeRequest(request: &request)
+    }
+
+}
+
+// MARK: - Private
+
+private extension NetworkSession {
+
+    func executeRequest(request: inout URLRequest) async throws(NetworkError) -> Data {
+        // Content type
+        let httpMethodSupportsBody = request.method.hasRequestBody
+        let httpMethodHasBody = (request.httpBody != nil)
+
+        if httpMethodSupportsBody && httpMethodHasBody { // assume we are always encoding data as JSON
+            request.setValue("application/json", forHTTPHeaderField: "Content-Type")
+        } else if httpMethodSupportsBody && !httpMethodHasBody { // supports body, but has parameters in query
+            request.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: "Content-Type")
+        } else if !httpMethodSupportsBody {
+            // do not set Content-Type
+        } else {
+            logger.logNetworkEvent(
+                message: "Cannot resolve Content-Type automatically",
+                level: .warning,
+                file: #file,
+                line: #line
+            )
+        }
+
+        // Session headers
+        sessionHeaders.forEach { header in
+            request.setValue(header.value, forHTTPHeaderField: header.name)
+        }
+
+        // Interceptors
+        try await interceptor.adapt(urlRequest: &request)
+
+        // Data task
+        let dataTask = session.dataTask(with: request)
+        let dataTaskProxy = proxyForTask(dataTask)
+        dataTask.resume()
+
+        // Request data + validation + retry (?)
+        do {
+            let data = try await dataTaskProxy.data()
+            closeProxyForTask(dataTask)
+            
+            let validator = DefaultValidationProvider()
+            let statusCode = (dataTask.response as? HTTPURLResponse)?.statusCode ?? -1
+            try validator.validate(statusCode: statusCode, data: data)
+            return data
+        } catch let networkError {
+            return try await retryRequest(request: &request, error: networkError)
+        }
+    }
+
+    func retryRequest(request: inout URLRequest, error networkError: NetworkError) async throws(NetworkError) -> Data {
+        let retryResult = try await interceptor.retry(urlRequest: &request, for: self, dueTo: networkError)
+
+        switch retryResult {
+        case .doNotRetry:
+            throw networkError
+
+        case .retryAfter(let timeInterval):
+            do {
+                try await Task.sleep(nanoseconds: UInt64(timeInterval * 10e9))
+            } catch {
+                throw URLError(.cancelled).asNetworkError()
+            }
+            fallthrough
+
+        case .retry:
+            return try await self.executeRequest(request: &request)
+        }
+    }
+    
+    @available(*, deprecated, message: "Unify parameter handling")
+    private func applyQueryParameters(to request: inout URLRequest, _ endpoint: any Endpoint, _ url: URL) {
+        if #available(iOS 16, macOS 13, *) {
+            request.url?.append(queryItems: endpoint.parameters?.queryItems() ?? [])
+        } else {
+            var urlComponents = URLComponents(url: url, resolvingAgainstBaseURL: false)
+            urlComponents?.queryItems?.append(contentsOf: endpoint.parameters?.queryItems() ?? [])
+            request.url = urlComponents?.url
+        }
+    }
+
+}
+
+// MARK: - Sample
+
+// The following code is a sample for testing the syntax.
+//private func sample() async {
+//    let session = NetworkSession(
+//        baseUrl: "https://api.sampleapis.com/",
+//        baseHeaders: [HTTPHeader("User-Agent: iOS app")],
+//        interceptor: CompositeInterceptor(interceptors: [
+//            AuthenticationInterceptor(authenticator: NoAuthenticator())
+//        ])
+//    )
+//
+//    do {
+//        let coffeeListA: String = try await session.request(endpoint: CoffeeEndpoint.hot)
+//        let coffeeListB: Data = try await session.get("/cards/list")
+//        
+//        
+//        let answers = Data()
+//        let data: Data = try await session.post("/coffee/survey/new")
+//        
+//        try await session.delete("/coffee/3")
+//        
+//        try await session.request(
+//            endpoint: at("/coffee/4")
+//                .method(.get)
+//                .header("Content-Type: application/xml")
+//                .body(data: answers)
+//        )
+//    } catch let error {
+//        assert(error is NetworkError)
 //    }
+//}
 //
-//    nonisolated public func hash(into hasher: inout Hasher) {
-//        hasher.combine(sessionId)
-//    }
+//enum CoffeeEndpoint: Endpoint {
 //
-//    // MARK: - ID
+//    case hot
 //
-//    nonisolated private let sessionId: UUID = UUID()
-//
-//    // MARK: - Properties
-//
-//    /// The provider responsible for managing the network session, ensuring it is created, resolved, and validated.
-//    public let sessionProvider: NetworkSessionProviding
-//
-//    /// The optional provider for resolving the base URL to be used in network requests.
-//    public let baseUrlProvider: BaseUrlProviding?
-//
-//    // MARK: - Initialization
-//
-//    /// Initializes the `NetworkSession` with an optional base URL provider and a session provider.
-//    ///
-//    /// - Parameters:
-//    ///   - baseUrlProvider: An optional provider for the base URL. Defaults to `nil`.
-//    ///   - sessionProvider: The session provider to be used. Defaults to `DefaultSessionProvider` with a default configuration.
-//    public init(
-//        baseUrlProvider: BaseUrlProviding? = nil,
-//        sessionProvider: NetworkSessionProviding = DefaultSessionProvider(configuration: .default())
-//    ) {
-//        self.baseUrlProvider = baseUrlProvider
-//        self.sessionProvider = sessionProvider
-//    }
-//
-//    /// Initializes the `NetworkSession` with an optional base URL provider and a network session configuration.
-//    ///
-//    /// - Parameters:
-//    ///   - baseUrl: An optional provider for the base URL. Defaults to `nil`.
-//    ///   - configuration: The configuration to be used for creating the session. Defaults to `.default`.
-//    public init(
-//        baseUrl: BaseUrlProviding? = nil,
-//        configuration: NetworkSessionConfiguration = .default(),
-//        logger: NetworkLogger? = nil
-//    ) {
-//        self.baseUrlProvider = baseUrl
-//        self.sessionProvider = DefaultSessionProvider(configuration: configuration, logger: logger)
-//    }
-//
-//    /// Initializes the `NetworkSession` with an optional base URL provider and an existing session.
-//    ///
-//    /// - Parameters:
-//    ///   - baseUrlProvider: An optional provider for the base URL. Defaults to `nil`.
-//    ///   - session: An existing session to be used by this provider.
-//    public init(
-//        baseUrlProvider: BaseUrlProviding? = nil,
-//        session: Alamofire.Session
-//    ) {
-//        self.baseUrlProvider = baseUrlProvider
-//        self.sessionProvider = DefaultSessionProvider(session: session)
-//    }
+//    var method: HTTPMethod { .get }
+//    var path: URLConvertible { "/coffee/hot" }
+//    var headers: HTTPHeaders? { nil }
+//    var parameters: EndpointParameters? { nil }
 //
 //}
 //
-//// MARK: - Request
 //
-//public extension OLD_NetworkSession {
-//
-//    /// Sends a network request to an endpoint using the resolved base URL and session.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - baseUrlProvider: An optional base URL provider. If `nil`, the default `baseUrlProvider` is used.
-//    ///   - validationProvider: The validation provider used to validate the response. Defaults to `DefaultValidationProvider`.
-//    /// - Returns: The decoded result of type `Result`.
-//    /// - Throws: A `Failure` error if validation or the request fails.
-//    func request<Result: Decodable & Sendable, Failure: Error>(
-//        endpoint: Endpoint,
-//        baseUrlProvider: BaseUrlProviding? = nil,
-//        validationProvider: any ValidationProviding<Failure> = DefaultValidationProvider()
-//    ) async throws(Failure) -> Result {
-//        return try await catchingFailure(validationProvider: validationProvider) {
-//            let resolvedBaseUrl = try await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//            let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//            return try await resolvedSession.request(
-//                try? endpoint.url(on: resolvedBaseUrl),
-//                method: endpoint.method,
-//                parameters: endpoint.parameters?.dictionary,
-//                encoding: endpoint.encoding,
-//                headers: endpoint.headers
-//            )
-//            .goodify(type: Result.self, validator: validationProvider)
-//            .value
-//        }
-//    }
-//
-//    /// Sends a raw network request and returns the response data.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - baseUrlProvider: An optional base URL provider. If `nil`, the default `baseUrlProvider` is used.
-//    ///   - validationProvider: The validation provider used to validate the response. Defaults to `DefaultValidationProvider`.
-//    /// - Returns: The raw response data.
-//    /// - Throws: A `Failure` error if validation or the request fails.
-//    func requestRaw<Failure: Error>(
-//        endpoint: Endpoint,
-//        baseUrlProvider: BaseUrlProviding? = nil,
-//        validationProvider: any ValidationProviding<Failure> = DefaultValidationProvider()
-//    ) async throws(Failure) -> Data {
-//        return try await catchingFailure(validationProvider: validationProvider) {
-//            let resolvedBaseUrl = try await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//            let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//            return try await resolvedSession.request(
-//                try? endpoint.url(on: resolvedBaseUrl),
-//                method: endpoint.method,
-//                parameters: endpoint.parameters?.dictionary,
-//                encoding: endpoint.encoding,
-//                headers: endpoint.headers
-//            )
-//            .serializingData()
-//            .value
-//        }
-//    }
-//
-//    /// Sends a request and returns an unprocessed `DataRequest` object.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - baseUrlProvider: An optional base URL provider. If `nil`, the default `baseUrlProvider` is used.
-//    /// - Returns: A `DataRequest` object representing the raw request.
-//    @_disfavoredOverload func request(endpoint: Endpoint, baseUrlProvider: BaseUrlProviding? = nil) async -> DataRequest {
-//        let resolvedBaseUrl = try? await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//        let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//        return resolvedSession.request(
-//            try? endpoint.url(on: resolvedBaseUrl ?? ""),
-//            method: endpoint.method,
-//            parameters: endpoint.parameters?.dictionary,
-//            encoding: endpoint.encoding,
-//            headers: endpoint.headers
-//        )
-//    }
-//
-//}
-//
-//// MARK: - Download
-//
-//public extension OLD_NetworkSession {
-//
-//    /// Creates a download request for the given `endpoint` and saves the result to the specified file.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - baseUrlProvider: An optional base URL provider. Defaults to `nil`.
-//    ///   - customFileName: The name of the file to which the downloaded content will be saved.
-//    /// - Returns: A `DownloadRequest` for the file download.
-//    /// - Throws: A `NetworkError` if the request fails.
-//    func download(endpoint: Endpoint, baseUrlProvider: BaseUrlProviding? = nil, customFileName: String) async throws(NetworkError) -> DownloadRequest {
-//        let resolvedBaseUrl = try await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//        let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//        let destination: DownloadRequest.Destination = { temporaryURL, _ in
-//            let directoryURLs = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-//            let url = directoryURLs.first?.appendingPathComponent(customFileName) ?? temporaryURL
-//
-//            return (url, [.removePreviousFile, .createIntermediateDirectories])
-//        }
-//
-//        return resolvedSession.download(
-//            try? endpoint.url(on: resolvedBaseUrl),
-//            method: endpoint.method,
-//            parameters: endpoint.parameters?.dictionary,
-//            encoding: endpoint.encoding,
-//            headers: endpoint.headers,
-//            to: destination
-//        )
-//    }
-//
-//}
-//
-//// MARK: - Upload
-//
-//public extension OLD_NetworkSession {
-//
-//    /// Uploads data to the specified `endpoint` using multipart form data.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - data: The data to be uploaded.
-//    ///   - fileHeader: The header to use for the uploaded file in the form data. Defaults to "file".
-//    ///   - filename: The name of the file to be uploaded.
-//    ///   - mimeType: The MIME type of the file.
-//    ///   - baseUrlProvider: An optional base URL provider. Defaults to `nil`.
-//    /// - Returns: An `UploadRequest` representing the upload.
-//    /// - Throws: A `NetworkError` if the upload fails.
-//    func uploadWithMultipart(
-//        endpoint: Endpoint,
-//        data: Data,
-//        fileHeader: String = "file",
-//        filename: String,
-//        mimeType: String,
-//        baseUrlProvider: BaseUrlProviding? = nil
-//    ) async throws(NetworkError) -> UploadRequest {
-//        let resolvedBaseUrl = try await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//        let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//        return resolvedSession.upload(
-//            multipartFormData: { formData in
-//                formData.append(data, withName: fileHeader, fileName: filename, mimeType: mimeType)
-//            },
-//            to: try? endpoint.url(on: resolvedBaseUrl),
-//            method: endpoint.method,
-//            headers: endpoint.headers
-//        )
-//    }
-//
-//    /// Uploads multipart form data to the specified `endpoint`.
-//    ///
-//    /// - Parameters:
-//    ///   - endpoint: The endpoint instance representing the URL, method, parameters, and headers.
-//    ///   - multipartFormData: The multipart form data to upload.
-//    ///   - baseUrlProvider: An optional base URL provider. Defaults to `nil`.
-//    /// - Returns: An `UploadRequest` representing the upload.
-//    /// - Throws: A `NetworkError` if the upload fails.
-//    func uploadWithMultipart(
-//        endpoint: Endpoint,
-//        multipartFormData: MultipartFormData,
-//        baseUrlProvider: BaseUrlProviding? = nil
-//    ) async throws(NetworkError) -> UploadRequest {
-//        let resolvedBaseUrl = try await resolveBaseUrl(baseUrlProvider: baseUrlProvider)
-//        let resolvedSession = await resolveSession(sessionProvider: sessionProvider)
-//
-//        return resolvedSession.upload(
-//            multipartFormData: multipartFormData,
-//            to: try? endpoint.url(on: resolvedBaseUrl),
-//            method: endpoint.method,
-//            headers: endpoint.headers
-//        )
-//    }
-//
-//}
-//
-//// MARK: - Internal
-//
-//extension OLD_NetworkSession {
-//
-//    /// Resolves the network session, creating a new one if necessary.
-//    ///
-//    /// - Parameter sessionProvider: The provider managing the session.
-//    /// - Returns: The resolved or newly created `Alamofire.Session`.
-//    func resolveSession(sessionProvider: NetworkSessionProviding) async -> Alamofire.Session {
-//        if await !sessionProvider.isSessionValid {
-//            await sessionProvider.makeSession()
-//        } else {
-//            await sessionProvider.resolveSession()
-//        }
-//    }
-//
-//    /// Resolves the base URL using the provided or default base URL provider.
-//    ///
-//    /// - Parameter baseUrlProvider: An optional base URL provider. If `nil`, the default `baseUrlProvider` is used.
-//    /// - Returns: The resolved base URL as a `String`.
-//    /// - Throws: A `NetworkError.invalidBaseURL` if the base URL cannot be resolved.
-//    func resolveBaseUrl(baseUrlProvider: BaseUrlProviding?) async throws(NetworkError) -> String {
-//        let baseUrlProvider = baseUrlProvider ?? self.baseUrlProvider
-//        guard let resolvedBaseUrl = await baseUrlProvider?.resolveBaseUrl() else {
-//            throw .invalidBaseURL
-//        }
-//        return resolvedBaseUrl
-//    }
-//
-//    /// Executes a closure while catching and transforming failures.
-//    ///
-//    /// - Parameters:
-//    ///   - validationProvider: The provider used to transform any errors.
-//    ///   - body: The closure to execute.
-//    /// - Returns: The result of type `Result`.
-//    /// - Throws: A transformed error if the closure fails.
-//    func catchingFailure<Result: Decodable & Sendable, Failure: Error>(
-//        validationProvider: any ValidationProviding<Failure>,
-//        body: () async throws -> Result
-//    ) async throws(Failure) -> Result {
-//        do {
-//            return try await body()
-//        } catch let networkError as NetworkError {
-//            throw validationProvider.transformError(networkError)
-//        } catch let error as AFError {
-//            if let underlyingError = error.underlyingError as? Failure {
-//                throw underlyingError
-//            } else if let underlyingError = error.underlyingError as? NetworkError {
-//                throw validationProvider.transformError(underlyingError)
-//            } else {
-//                throw validationProvider.transformError(NetworkError.sessionError)
-//            }
-//        } catch {
-//            throw validationProvider.transformError(NetworkError.sessionError)
-//        }
-//    }
-//
+//enum Endpoints {
+//    static let hotCoffee = "/coffee/hot"
 //}
