--- Tests/GoodNetworkingTests/Resources/ArrayNil.json
@@ -1,3 +1,0 @@
-{
-    "ids": []
-}
