--- Tests/GoodNetworkingTests/Resources/EmptyElement.json
@@ -1,17 +1,0 @@
-{
-    "ids":
-    [
-        {
-            "id": 1
-        },
-        {
-            "id": 2
-        },
-        {
-
-        },
-        {
-            "id": 3
-        }
-    ]
-}
