--- Tests/GoodNetworkingTests/Resources/IsoDate.json
@@ -1,3 +1,0 @@
-{
-    "date": "2022-11-19T01:29:34+01:00"
-}
