--- Tests/GoodNetworkingTests/Resources/MilisecondsDate.json
@@ -1,3 +1,0 @@
-{
-    "date": 1678798320209
-}
